ADDING NEW LOCATIONS (requested by John) today, i add some new locations to the website. -uk -russia -germany -france
Right Now I finished the Russia(with the help of our little black friend by giving me the image of the head chief), which was a simple addition. it was just a long time that havent added a new location that I had to refer to the old code. I will add the username and password after im done with this side of the project.
After im done with all the other locations I will add the links that will redeitc the player to the actual website. to make it even more belivabal.